[FIX] On macOS, exit from main_depthMapEstimation.cpp without running destructors or cleaning the stack
On macOS, some logic issue in internal Objective-C code (might be caused by OpenMP) causes an AutoreleasePool to be released twice. The backtrace is highly unspecific and does not indicate any issue related to the AliceVision code. It can be avoided by calling std::_Exit(0) before exiting normally. This avoids a SIGSEGV at the very end of the program.

Revert if root cause is fixed.

Signed-off-by: Philipp Remy <philipp.remy@law-school.de>